requirements.txt
tensorflow==2.3.0
scikit-learn==0.23.1
lightgbm==2.3.1
XGBoost==1.1.1
yara-python==4.0.2
lief==0.10.1
capstone==4.0.2
pillow==7.2.0
tqdm==4.48.2
numpy==1.18.5
pandas==1.1.0
matplotlib==3.3.0